move play scene to separate file
